feat: simplify MCP host SDK ergonomics
