+++ examples/sdk/README.md
@@ -1,0 +1,14 @@
+# SDK Examples
+
+These examples move from the smallest interactive host toward more customized
+Heddle SDK usage.
+
+## 01 Interactive Chat
+
+```bash
+yarn example:sdk:interactive
+```
+
+This starts a persisted conversation loop with default text output and local
+commands. It is the recommended first file to copy when trying Heddle as a
+framework.
