+++ package.json
@@ -74,6 +74,7 @@
     "chat:dev:anthropic": "ANTHROPIC_API_KEY=\"$PERSONAL_ANTHROPIC_API_KEY\" tsx --no-cache src/cli-v2/main.ts chat",
     "chat:dev:v2:anthropic": "ANTHROPIC_API_KEY=\"$PERSONAL_ANTHROPIC_API_KEY\" tsx --no-cache src/cli-v2/main.ts chat-v2",
     "example:repo-investigator": "tsx --no-cache examples/repo-investigator.ts",
+    "example:sdk:interactive": "tsx --no-cache examples/sdk/01-interactive-chat.ts",
     "example:conversation-engine": "tsx --no-cache examples/conversation-engine.ts",
     "example:programmatic": "tsx --no-cache examples/programmatic-loop.ts",
     "example:heartbeat": "tsx --no-cache examples/heartbeat.ts",
