+++ src/core/chat/engine/conversation-cli/index.ts
@@ -1,0 +1,2 @@
+export { ConversationCliRunnerService, runConversationCli } from './service.js';
+export type { ConversationCliRunnerOptions } from './types.js';
